Added tool for dumping shader program binaries. Messed with SLZLighting and other includes to try to optimize LitMAS's register usage. Fixed some unity 6 API changes in the test scripts. Other old changes I don't remember.

+++ ShaderLibrary/VolumetricCore.hlsl
@@ -134,16 +134,16 @@ half4 VolumetricsSurf(half4 color, float3 positionWS, int surfaceType) {
 
 float4 _MipFogParameters = float4(0,5,0.5,0);
 
-float EvaluateMonochromaticSHL2(float3 normal)
+half EvaluateMonochromaticSHL2(half3 normal)
 {
     // Monochromatic SH evaluation using the coefficients array
-    float shValue = _SHMonoCoefficients[0] +                             // L0 term (constant)
+	half shValue = _SHMonoCoefficients[0] + // L0 term (constant)
                     normal.y * _SHMonoCoefficients[1] +                  // L1 Y term (gradient)
                     normal.z * _SHMonoCoefficients[2] +                  // L1 Z term
                     normal.x * _SHMonoCoefficients[3] +                  // L1 X term
                     normal.x * normal.y * _SHMonoCoefficients[4] +       // L2 XY term
                     normal.y * normal.z * _SHMonoCoefficients[5] +       // L2 YZ term
-                    (3.0 * normal.z * normal.z - 1.0) * _SHMonoCoefficients[6] +  // L2 Z² term
+                    (half(3.0) * normal.z * normal.z - half(1.0)) * _SHMonoCoefficients[6] + // L2 Z² term
                     normal.x * normal.z * _SHMonoCoefficients[7] +       // L2 XZ term
                     (normal.x * normal.x - normal.y * normal.y) * _SHMonoCoefficients[8];  // L2 X² - Y² term
